[WIP/Update] Handle %x with min_field_width, precision and some flags.
Almost all the case is handled current implementation, but for one case.
When 0 is given to the precision, the original printf is print nothing,
but mine is print single character, fix it before moveon to next task.

Also, I suppose to create print logic full scratch. Due to complexity of
printing padding character etc.